Split code into multiple files, fix tests and doc tests
