+++ Cargo.toml
@@ -20,7 +20,6 @@ serde_json = ">= 1.0.0"
 serde_yaml = ">= 0.8.0"
 uuid = { version = ">= 0.21.0", features = ["v4"] }
 dirs = ">= 2.0.0"
-base64 = ">= 0.10.0"
 chrono = ">= 0.4.0"
 tokio = { version = ">= 1.0.1", features = ["fs"] }
 hex = ">= 0.4.0"
@@ -29,6 +28,7 @@ http = ">= 1.0.0"
 tower = { version = ">= 0.4.13", optional = true }
 axum = { version = ">= 0.7.2", optional = true }
 futures-core = { version = ">= 0.3.25", optional = true }
+base64 = { version = ">= 0.10.0", optional = true }
 thiserror = ">= 1.0.37"
 
 [dependencies.mauth-core]
@@ -40,4 +40,4 @@ version = "0.5"
 tokio = { version = ">= 1.0.1", features = ["rt-multi-thread", "macros"] }
 
 [features]
-axum-service = ["tower", "futures-core", "axum"]
+axum-service = ["tower", "futures-core", "axum", "base64"]
