+++ README.md
@@ -13,19 +13,11 @@ approval for the full stack used through the Architecture and Security groups.
 let mauth_info = MAuthInfo::from_default_file().unwrap();
 let client = Client::new();
 let uri: Url = "https://www.example.com/".parse().unwrap();
-let (body, body_digest) = MAuthInfo::build_body_with_digest("".to_string());
 let mut req = Request::new(Method::GET, uri);
-*req.body_mut() = Some(body);
-mauth_info.sign_request(&mut req, &body_digest);
+mauth_info.sign_request(&mut req, &[]);
 match client.execute(req).await {
     Err(err) => println!("Got error {}", err),
-    Ok(response) => match mauth_info.validate_response(response).await {
-        Ok(resp_body) => println!(
-            "Got validated response with body {}",
-            &String::from_utf8(resp_body).unwrap()
-        ),
-        Err(err) => println!("Error validating response: {:?}", err),
-    }
+    Ok(response) => println!("Got validated response with body {}", response.text().await.unwrap()),
 }
 ```
